jones_cup_api/pyproject.toml
[project]
name = "template-fastapi-python"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "google-cloud-storage (>=2.16.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

